+++ forui_hooks/lib/src/date_field_controller_hook.dart
@@ -12,56 +12,27 @@ String? _defaultValidator(DateTime? _) => null;
 /// [validator] returns an error string to display if the input is invalid, or null otherwise.
 /// It is also used to determine whether a date in a calendar is selectable.
 /// Defaults to always returning null.
-///
-/// [truncateAndStripTimezone] determines whether the controller should truncate and convert all given [DateTime]s to
-/// dates in UTC timezone. Defaults to true.
-///
-/// ```dart
-/// DateTime truncateAndStripTimezone(DateTime date) => DateTime.utc(date.year, date.month, date.day);
-/// ```
-///
-/// [truncateAndStripTimezone] should be set to false if you can guarantee that all dates are in UTC timezone (with
-/// the help of a 3rd party library), which will improve performance. **Warning:** Giving a [DateTime] in local
-/// timezone or with a time component when [truncateAndStripTimezone] is false is undefined behavior.
-///
-/// ## Contract
-/// Throws [AssertionError] if [initialDate] is not in UTC timezone and [truncateAndStripTimezone] is false.
 FDateFieldController useFDateFieldController({
-  TickerProvider? vsync,
-  DateTime? initialDate,
+  DateTime? initial,
   FormFieldValidator<DateTime> validator = _defaultValidator,
-  bool truncateAndStripTimezone = true,
-  FPopoverMotion popoverMotion = const FPopoverMotion(),
   List<Object?>? keys,
 }) => use(
   _DateFieldHook(
-    vsync: vsync ??= useSingleTickerProvider(keys: keys),
-    initialDate: initialDate,
+    initial: initial,
     validator: validator,
-    truncateAndStripTimezone: truncateAndStripTimezone,
-    popoverMotion: popoverMotion,
-    debugLabel: 'useFDatePickerController',
     keys: keys,
   ),
 );
 
 class _DateFieldHook extends Hook<FDateFieldController> {
-  final TickerProvider vsync;
-  final DateTime? initialDate;
+  final DateTime? initial;
   final FormFieldValidator<DateTime> validator;
-  final bool truncateAndStripTimezone;
-  final FPopoverMotion popoverMotion;
-  final String _debugLabel;
 
   const _DateFieldHook({
-    required this.vsync,
-    required this.initialDate,
+    required this.initial,
     required this.validator,
-    required this.truncateAndStripTimezone,
-    required this.popoverMotion,
-    required String debugLabel,
     super.keys,
-  }) : _debugLabel = debugLabel;
+  });
 
   @override
   _DateFieldHookState createState() => .new();
@@ -70,23 +41,15 @@ class _DateFieldHook extends Hook<FDateFieldController> {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
-      ..add(DiagnosticsProperty('vsync', vsync))
-      ..add(DiagnosticsProperty('initialDate', initialDate))
-      ..add(ObjectFlagProperty.has('validator', validator))
-      ..add(
-        FlagProperty('truncateAndStripTimezone', value: truncateAndStripTimezone, ifTrue: 'truncateAndStripTimezone'),
-      )
-      ..add(DiagnosticsProperty('popoverMotion', popoverMotion));
+      ..add(DiagnosticsProperty('initial', initial))
+      ..add(ObjectFlagProperty.has('validator', validator));
   }
 }
 
 class _DateFieldHookState extends HookState<FDateFieldController, _DateFieldHook> {
   late final _controller = FDateFieldController(
-    vsync: hook.vsync,
-    initialDate: hook.initialDate,
+    initial: hook.initial,
     validator: hook.validator,
-    truncateAndStripTimezone: hook.truncateAndStripTimezone,
-    popoverMotion: hook.popoverMotion,
   );
 
   @override
@@ -99,5 +62,5 @@ class _DateFieldHookState extends HookState<FDateFieldController, _DateFieldHook
   bool get debugHasShortDescription => false;
 
   @override
-  String get debugLabel => hook._debugLabel;
+  String get debugLabel => 'useFDateFieldController';
 }